+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 v1.03.15
 =========
 
+* Show the Order time in the Recent Orders table of the Admin Dashboard.
 * Remove newly sold-out Products from Carts when each Order is placed instead
   of on an hourly basis.
 * Add a hover popup to the Admin Dashboard graphs showing the Date & Sales